update react and dependencies
