+++ frontend/src/Cells.js
@@ -1,20 +1,20 @@
 import React, { useEffect, useState } from 'react';
-import Typography from '@material-ui/core/Typography';
-import TextField from '@material-ui/core/TextField';
-import { makeStyles } from '@material-ui/core/styles';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import Paper from '@material-ui/core/Paper';
-import Tooltip from '@material-ui/core/Tooltip';
-import AddIcon from '@material-ui/icons/Add';
-import IconButton from '@material-ui/core/IconButton';
-import TableCell from '@material-ui/core/TableCell';
-import Container from '@material-ui/core/Container';
-import SaveIcon from '@material-ui/icons/Save';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import DeleteIcon from '@material-ui/icons/Delete';
-import TableRow from '@material-ui/core/TableRow';
+import Typography from '@mui/material/Typography';
+import TextField from '@mui/material/TextField';
+import { makeStyles } from '@mui/styles'
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import Paper from '@mui/material/Paper';
+import Tooltip from '@mui/material/Tooltip';
+import AddIcon from '@mui/icons-material/Add';
+import IconButton from '@mui/material/IconButton';
+import TableCell from '@mui/material/TableCell';
+import Container from '@mui/material/Container';
+import SaveIcon from '@mui/icons-material/Save';
+import TableContainer from '@mui/material/TableContainer';
+import TableHead from '@mui/material/TableHead';
+import DeleteIcon from '@mui/icons-material/Delete';
+import TableRow from '@mui/material/TableRow';
 import { 
   cells, add_cell, 
   update_cell, delete_cell
