+++ frontend/src/FormComponents.js
@@ -1,15 +1,15 @@
 import React from "react";
-import TextField from '@material-ui/core/TextField';
-import { makeStyles } from '@material-ui/core/styles';
-import InputLabel from '@material-ui/core/InputLabel';
-import MenuItem from '@material-ui/core/MenuItem';
-import Autocomplete from '@material-ui/lab/Autocomplete';
-import FormControl from '@material-ui/core/FormControl';
-import Select from '@material-ui/core/Select';
+import TextField from '@mui/material/TextField';
+import { makeStyles } from '@mui/styles'
+import InputLabel from '@mui/material/InputLabel';
+import MenuItem from '@mui/material/MenuItem';
+import Autocomplete from '@mui/lab/Autocomplete';
+import FormControl from '@mui/material/FormControl';
+import Select from '@mui/material/Select';
 import { Controller  } from "react-hook-form";
 //import DateFnsUtils from '@date-io/date-fns'; // choose your lib
-//import { DateTimePicker } from '@material-ui/pickers';
-import Chip from '@material-ui/core/Chip';
+//import { DateTimePicker } from '@mui/pickers';
+import Chip from '@mui/material/Chip';
 
 const useStyles = makeStyles((theme) => ({
   container: {
